# pygmt.grdvolume: Refactor to store output in virtual files instead of temporary files