--- a/pygmt/tests/test_grdvolume.py
+++ b/pygmt/tests/test_grdvolume.py
@@ -2,15 +2,11 @@
 Test pygmt.grdvolume.
 """
 
-from pathlib import Path
-
 import numpy as np
 import numpy.testing as npt
 import pandas as pd
 import pytest
 from pygmt import grdvolume
-from pygmt.exceptions import GMTInvalidInput
-from pygmt.helpers import GMTTempFile
 from pygmt.helpers.testing import load_static_earth_relief
 
 
@@ -20,14 +16,6 @@ def fixture_grid():
     Load the grid data from the sample earth_relief file.
     """
     return load_static_earth_relief()
-
-
-@pytest.fixture(scope="module", name="region")
-def fixture_region():
-    """
-    Set the data region for the tests.
-    """
-    return [-53, -50, -22, -20]
 
 
 @pytest.fixture(scope="module", name="data")
@@ -47,57 +35,15 @@ def fixture_data():
     return data
 
 
-def test_grdvolume_format(grid, region):
-    """
-    Test that correct formats are returned.
-    """
-    grdvolume_default = grdvolume(grid=grid, region=region)
-    assert isinstance(grdvolume_default, pd.DataFrame)
-    grdvolume_array = grdvolume(grid=grid, output_type="numpy", region=region)
-    assert isinstance(grdvolume_array, np.ndarray)
-    grdvolume_df = grdvolume(grid=grid, output_type="pandas", region=region)
-    assert isinstance(grdvolume_df, pd.DataFrame)
-
-
-def test_grdvolume_invalid_format(grid):
-    """
-    Test that grdvolume fails with incorrect output_type argument.
-    """
-    with pytest.raises(GMTInvalidInput):
-        grdvolume(grid=grid, output_type=1)
-
-
-def test_grdvolume_no_outfile(grid):
-    """
-    Test that grdvolume fails when output_type set to 'file' but no outfile is
-    specified.
-    """
-    with pytest.raises(GMTInvalidInput):
-        grdvolume(grid=grid, output_type="file")
-
-
 @pytest.mark.benchmark
-def test_grdvolume_no_outgrid(grid, data, region):
+def test_grdvolume(grid, data):
     """
-    Test the expected output of grdvolume with no output file set.
+    Test the basic functionality of grdvolume.
     """
     test_output = grdvolume(
-        grid=grid, contour=[200, 400, 50], output_type="numpy", region=region
+        grid=grid,
+        contour=[200, 400, 50],
+        region=[-53, -50, -22, -20],
     )
+    assert isinstance(test_output, pd.DataFrame)
     npt.assert_allclose(test_output, data)
-
-
-def test_grdvolume_outgrid(grid, region):
-    """
-    Test the expected output of grdvolume with an output file set.
-    """
-    with GMTTempFile(suffix=".csv") as tmpfile:
-        result = grdvolume(
-            grid=grid,
-            contour=[200, 400, 50],
-            output_type="file",
-            outfile=tmpfile.name,
-            region=region,
-        )
-        assert result is None  # return value is None
-        assert Path(tmpfile.name).stat().st_size > 0  # check that outfile exists