Address #1318 and #2730.

**Preview**: https://pygmt-dev--3102.org.readthedocs.build/en/3102/api/generated/pygmt.grdvolume.html#pygmt.grdvolume

Benchmarks:

```
>>> from pygmt import grdvolume
>>> for i in [50, 25, 10, 5, 2, 1, 0.1]:
...     %timeit df = grdvolume("@earth_relief_01d_g", contour=[200, 400, i])
...
```
This branch:
```
12.7 ms ± 42 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
15.8 ms ± 78 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
25.1 ms ± 149 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
40.8 ms ± 244 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
87.2 ms ± 209 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
165 ms ± 227 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
1.57 s ± 4.11 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

Main branch:
```
10.3 ms ± 39.6 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
13.4 ms ± 23.3 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
22.4 ms ± 29.2 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
37.6 ms ± 43.2 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
83 ms ± 136 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
159 ms ± 225 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
1.52 s ± 1.12 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```
This branch is slightly slower than the main branch, mainly because:

1. this module is computational-intensive, so most of the time are spent in computation, rather than I/O
2. For most real cases, the number of specified contours is small (e.g., 2000 for the last run), so using tempfile or virtualfile should have very close performance.